# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentQuery": {
        "properties": {
          "input": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "title": "AgentQuery",
        "type": "object"
      },
...
      "AgentSuccessResponse": {
        "description": "Standard envelope for a successful non-streaming agent response.",
        "properties": {
          "result": {
            "title": "Result",
            "type": "string"
          }
        },
        "required": [
          "result"
        ],
        "title": "AgentSuccessResponse",
        "type": "object"
      },
...
      "PendingActionResponse": {
        "description": "Immutable, owner-bound external action awaiting or reflecting approval.",
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/PendingActionType"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "attempt_count": {
            "default": 0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "payload": {
            "title": "Payload",
            "type": "object"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "reconciliation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconciliation Reason"
          },
          "reconciliation_required_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconciliation Required At"
          },
          "result": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/PendingActionStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action_type",
          "payload",
          "payload_hash",
          "summary",
          "status",
          "created_at",
          "expires_at"
        ],
        "title": "PendingActionResponse",
        "type": "object"
      },
      "PendingActionStatus": {
        "description": "Persisted pending-action lifecycle states.",
        "enum": [
          "pending",
          "processing",
          "succeeded",
          "failed",
          "rejected",
          "expired",
          "reconciliation_required"
        ],
        "title": "PendingActionStatus",
        "type": "string"
      },
      "PendingActionType": {
        "description": "Provider effects that are always represented by a pending action.",
        "enum": [
          "send_email",
          "send_reply",
          "create_event",
          "create_task"
        ],
        "title": "PendingActionType",
        "type": "string"
      },
      "PublicErrorResponse": {
        "description": "Stable safe HTTP error envelope shared by all routes.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hexel Onebox Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/actions/{action_id}": {
      "get": {
        "operationId": "get_pending_action_endpoint_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Pending Action Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
    "/actions/{action_id}/approve": {
      "post": {
        "description": "Approve and execute exactly one immutable action owned by this JWT user.",
        "operationId": "approve_pending_action_endpoint_actions__action_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Pending Action Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
    "/actions/{action_id}/reconcile": {
      "post": {
        "description": "Reconcile an ambiguous write using deterministic provider markers only.",
        "operationId": "reconcile_pending_action_endpoint_actions__action_id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reconcile Pending Action Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
    "/actions/{action_id}/reject": {
      "post": {
        "operationId": "reject_pending_action_endpoint_actions__action_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Pending Action Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
...
    "/executive/": {
      "post": {
        "operationId": "invoke_executive_agent_endpoint_executive__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invoke Executive Agent Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
    "/generate-content/": {
      "post": {
        "operationId": "invoke_general_agent_endpoint_generate_content__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invoke General Agent Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
    "/generate-stream/": {
      "post": {
        "description": "Stream structured, user-safe agent events as JSON SSE data frames.",
        "operationId": "invoke_general_agent_stream_endpoint_generate_stream__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"event\":\"token\",\"content\":\"Hello\"}\\n\\n",
                "schema": {
                  "format": "event-stream",
                  "type": "string"
                }
              }
            },
            "description": "A text/event-stream response. Each non-comment data frame is JSON matching AgentStreamEvent."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invoke General Agent Stream Endpoint",
        "tags": [
          "AI Agents"
        ]
      }
    },
...
  }
}